Fix: 매매시스템 Wallet, STOS 연동

+++ spring/src/main/java/com/hana1piece/wallet/model/vo/WalletVO.java
@@ -8,6 +8,6 @@ public class WalletVO {
     private String memberId;
     private String accountNumber;
     private String password;
-    private int balance;
+    private long balance;
     private String openingDate;
 }
